// src/types/user/user.types.ts
export type User = {
  CODEMP: string;
  CODUSU: string;
  CODVEND: string;
  NOMEUSU: string;
  codemp: string;
  codusu: string;
  codvend: string;
  erro: string;
  nomeusu: string;
  ACESSOVISUALCAB: string;
  CODCENCUS: string;
  SEQVISITA: number[];
  vendedores: {
    CODVEND: string;
    APELIDO: string;
  }[];
};